+++ examples/get_ble_data.py
@@ -3,6 +3,9 @@ Get all BLE device broadcasts
 '''
 
 from ruuvitag_sensor.ble_communication import BleCommunicationNix
+from ruuvitag_sensor.log import configureLog
+
+configureLog(True)
 
 ble = BleCommunicationNix()