+++ ruuvitag_sensor/__main__.py
@@ -1,23 +1,26 @@
 import sys
 import argparse
+import logging
 
 import ruuvitag_sensor
 import ruuvitag_sensor.log
 from ruuvi import RuuviTagSensor
-from log import logger  # pylint: disable=E0611
+from log import configureLog  # pylint: disable=E0611
+
+configureLog(True)
+
+log = logging.getLogger('ruuvitag_sensor')
 
 
 def my_excepthook(exctype, value, traceback):
     sys.__excepthook__(exctype, value, traceback)
 
     if not issubclass(exctype, KeyboardInterrupt):
-        logger.critical(value)
+        log.critical(value)
 
 sys.excepthook = my_excepthook
 
 if __name__ == '__main__':
-    ruuvitag_sensor.log.enableConsole()
-
     parser = argparse.ArgumentParser()
     parser.add_argument('-g', '--get', dest='mac_address', help='Get data')
     parser.add_argument('-f', '--find', action='store_true',
@@ -33,13 +36,13 @@ if __name__ == '__main__':
     if args.mac_address:
         sensor = RuuviTagSensor(args.mac_address)
         state = sensor.update()
-        logger.info(state)
+        log.info(state)
     elif args.find_action:
         RuuviTagSensor.find_ruuvitags()
     elif args.latest_action:
         datas = RuuviTagSensor.get_data_for_sensors()
-        logger.info(datas)
+        log.info(datas)
     elif args.stream_action:
-        RuuviTagSensor.get_datas(lambda x: logger.info('%s - %s' % (x[0], x[1])))
+        RuuviTagSensor.get_datas(lambda x: log.info('%s - %s' % (x[0], x[1])))
     else:
         parser.print_usage()