+++ ruuvitag_sensor/ble_communication.py
@@ -2,8 +2,10 @@ import abc
 import subprocess
 import sys
 import os
+import logging
 
-from ruuvitag_sensor.log import logger  # pylint: disable=E0611
+log = logging.getLogger('ruuvitag_sensor')
+
 
 # Eddystone Protocol specification
 # https://github.com/google/eddystone/blob/master/protocol-specification.md
@@ -50,7 +52,7 @@ class BleCommunicationNix(BleCommunication):
 
     @staticmethod
     def start():
-        logger.info('Start receiving broadcasts')
+        log.info('Start receiving broadcasts')
         DEVNULL = subprocess.DEVNULL if sys.version_info >= (3, 3) else open(os.devnull, 'wb')
 
         subprocess.call('sudo hciconfig hci0 reset', shell=True, stdout=DEVNULL)
@@ -63,7 +65,7 @@ class BleCommunicationNix(BleCommunication):
         # import psutil here so as long as all implementations are in the same file, all will work
         import psutil
 
-        logger.info('Stop receiving broadcasts')
+        log.info('Stop receiving broadcasts')
 
         def kill_child_processes(parent_pid):
             try:
@@ -96,7 +98,7 @@ class BleCommunicationNix(BleCommunication):
         except KeyboardInterrupt as ex:
             return
         except Exception as ex:
-            logger.info(ex)
+            log.info(ex)
             return
 
     @staticmethod
@@ -125,7 +127,7 @@ class BleCommunicationNix(BleCommunication):
         data_iter = BleCommunicationNix.get_datas()
         for data in data_iter:
             if mac == data[0]:
-                logger.info('Data found')
+                log.info('Data found')
                 data_iter.send(StopIteration)
                 data = data[1]
                 break